Clean up data model. Make migrations more cohesive.
Got rid of the distinction between id and handle, made some fields
required that should have been originally. Fixed issues where price
was sometimes a number and sometimes a string. Made price "starting At"
with the minimum of variants, rather than just the value of the first
variant.

There are some pieces of data that I just didn't know what to do with.
There's stuff about selected variants that seem to be realted to the
unrelated to the unimplemented filter features. I mostly just left
the data commented out. A future contributor can take that across the
finish line.